Potenciometer brushless control
This code may be used in order to calibrate the ESC (electronic speed controller) or simply to control a brushless motor. 
If you want to calibrate yours you'll have to power off the ESC and then upload the code to your Arduino. Once it's done, turn the potentiometer to it's higher value (which would mean maximum speed) and then power on your ESC. You'll hear a "beep" from it, wait 2 seconds and turn the potentiometer to it's lower value, you'll hear two "beeps" this time, and it will mean that your ESC have been calibrated correctly.